Extract sender to a separate class
