+++ vaadin-dev-server/src/main/java/com/vaadin/base/devserver/stats/StatisticsStorage.java
@@ -22,29 +22,18 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiConsumer;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.vaadin.flow.server.Command;
 
-import org.apache.http.HttpResponse;
-import org.apache.http.HttpStatus;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.client.HttpClientBuilder;
-import org.apache.http.util.EntityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Development more usage statistic storage and methods for loading, saving and
- * sending the data.
+ * Development more usage statistic storage and methods for updating the data.
  */
 public class StatisticsStorage {
 
-    private static final String FAILED_TO_READ = "Failed to read ";
     private String projectId;
-    private String usageReportingUrl;
     private File usageStatisticsFile;
 
     /**
@@ -52,74 +41,6 @@ public class StatisticsStorage {
      */
     public StatisticsStorage() {
         // Intentionally empty
-    }
-
-    /**
-     * Get interval that is between {@link StatisticsConstants#TIME_SEC_12H} and
-     * {@link StatisticsConstants#TIME_SEC_30D}
-     *
-     * @param intervalSec
-     *            Interval to normalize
-     * @return <code>interval</code> if inside valid range.
-     */
-    private static long normalizeInterval(long intervalSec) {
-        if (intervalSec < StatisticsConstants.TIME_SEC_12H) {
-            return StatisticsConstants.TIME_SEC_12H;
-        }
-        return Math.min(intervalSec, StatisticsConstants.TIME_SEC_30D);
-    }
-
-    /**
-     * Posts given Json data to a URL.
-     * <p>
-     * Updates <code>FIELD_LAST_STATUS</code>.
-     *
-     * @param postUrl
-     *            URL to post data to.
-     * @param data
-     *            Json data to send
-     * @return Response or <code>data</code> if the data was not successfully
-     *         sent.
-     */
-    private static ObjectNode postData(String postUrl, JsonNode data) {
-        ObjectNode result;
-        try {
-            HttpPost post = new HttpPost(postUrl);
-            post.addHeader("Content-Type", "application/json");
-            post.setEntity(new StringEntity(
-                    JsonHelpers.getJsonMapper().writeValueAsString(data)));
-
-            HttpClient client = HttpClientBuilder.create().build();
-            HttpResponse response = client.execute(post);
-            String responseStatus = response.getStatusLine().getStatusCode()
-                    + ": " + response.getStatusLine().getReasonPhrase();
-            JsonNode jsonResponse = null;
-            if (response.getStatusLine().getStatusCode() == HttpStatus.SC_OK) {
-                String responseString = EntityUtils
-                        .toString(response.getEntity());
-                jsonResponse = JsonHelpers.getJsonMapper()
-                        .readTree(responseString);
-            }
-
-            if (jsonResponse != null && jsonResponse.isObject()) {
-                result = (ObjectNode) jsonResponse;
-            } else {
-                // Default response in case of any problems
-                result = JsonHelpers.getJsonMapper().createObjectNode();
-            }
-            // Update the status and return the results
-            result.put(StatisticsConstants.FIELD_LAST_STATUS, responseStatus);
-            return result;
-
-        } catch (IOException e) {
-            getLogger().debug("Failed to send statistics.", e);
-        }
-
-        // Fallback
-        result = JsonHelpers.getJsonMapper().createObjectNode();
-        result.put(StatisticsConstants.FIELD_LAST_STATUS,
-                StatisticsConstants.INVALID_SERVER_RESPONSE);
-        return result;
     }
 
     private static Logger getLogger() {
@@ -149,184 +70,6 @@ public class StatisticsStorage {
     }
 
     /**
-     * Get the remote reporting URL.
-     *
-     * @return Returns {@link StatisticsConstants#USAGE_REPORT_URL} by default.
-     */
-    String getUsageReportingUrl() {
-        return usageReportingUrl == null ? StatisticsConstants.USAGE_REPORT_URL
-                : usageReportingUrl;
-    }
-
-    /**
-     * Set the remote reporting URL.
-     *
-     * If not set, <code>StatisticsConstants.USAGE_REPORT_URL</code> is used.
-     *
-     * @param reportingUrl
-     *            Set the reporting URL.
-     */
-    void setUsageReportingUrl(String reportingUrl) {
-        this.usageReportingUrl = reportingUrl;
-    }
-
-    /**
-     * Helper to update client data in current project.
-     *
-     * @param clientData
-     *            Json data received from client.
-     */
-    void updateProjectTelemetryData(JsonNode clientData) {
-        try {
-            if (clientData != null && clientData.isObject()) {
-                update((global, project) -> {
-                    clientData.fields().forEachRemaining(
-                            e -> project.setValue(e.getKey(), e.getValue()));
-                });
-            }
-        } catch (Exception e) {
-            getLogger().debug("Failed to update client telemetry data", e);
-        }
-    }
-
-    /**
-     * Send current statistics to given reporting URL.
-     * <p>
-     * Reads the current data and posts it to given URL. Updates or replaces the
-     * local data according to the response.
-     *
-     * @see #postData(String, JsonNode)
-     */
-    String sendCurrentStatistics(ObjectNode json) {
-        // Post copy of the current data
-        AtomicReference<String> message = new AtomicReference<>(null);
-        JsonNode response = postData(getUsageReportingUrl(), json.deepCopy());
-
-        // Update the last sent time
-        // If the last send was successful we clear the project data
-        if (response.isObject()
-                && response.has(StatisticsConstants.FIELD_LAST_STATUS)) {
-            update((global, project) -> {
-
-                global.setValue(StatisticsConstants.FIELD_LAST_SENT,
-                        System.currentTimeMillis());
-                global.setValue(StatisticsConstants.FIELD_LAST_STATUS, response
-                        .get(StatisticsConstants.FIELD_LAST_STATUS).asText());
-
-                // Use different interval, if requested in response or default
-                // to 24H
-                if (response.has(StatisticsConstants.FIELD_SEND_INTERVAL)
-                        && response.get(StatisticsConstants.FIELD_SEND_INTERVAL)
-                                .isNumber()) {
-                    global.setValue(StatisticsConstants.FIELD_SEND_INTERVAL,
-                            normalizeInterval(response.get(
-                                    StatisticsConstants.FIELD_SEND_INTERVAL)
-                                    .asLong()));
-                } else {
-                    global.setValue(StatisticsConstants.FIELD_SEND_INTERVAL,
-                            StatisticsConstants.TIME_SEC_24H);
-                }
-
-                // Update the server message
-                if (response.has(StatisticsConstants.FIELD_SERVER_MESSAGE)
-                        && response
-                                .get(StatisticsConstants.FIELD_SERVER_MESSAGE)
-                                .isTextual()) {
-                    String msg = response
-                            .get(StatisticsConstants.FIELD_SERVER_MESSAGE)
-                            .asText();
-                    global.setValue(StatisticsConstants.FIELD_SERVER_MESSAGE,
-                            msg);
-                    message.set(msg);
-                }
-
-                // If data was sent ok, clear the existing project data
-                if (response.get(StatisticsConstants.FIELD_LAST_STATUS).asText()
-                        .startsWith("200:")) {
-                    global.setValue(StatisticsConstants.FIELD_PROJECTS,
-                            JsonHelpers.getJsonMapper().createArrayNode());
-                }
-            });
-        }
-
-        return message.get();
-    }
-
-    /**
-     * Check the Interval has elapsed.
-     * <p>
-     * Uses <code>System.currentTimeMillis</code> as time source.
-     *
-     * @return true if enough time has passed since the last send attempt.
-     * @see #getLastSendTime()
-     * @see #getInterval()
-     */
-    boolean isIntervalElapsed(ObjectNode json) {
-        long now = System.currentTimeMillis();
-        long lastSend = getLastSendTime(json);
-        long interval = getInterval(json);
-        return lastSend + interval * 1000 < now;
-    }
-
-    /**
-     * Reads the statistics update interval.
-     *
-     * @return Time interval in seconds.
-     *         {@link StatisticsConstants#TIME_SEC_24H} in minumun and
-     *         {@link StatisticsConstants#TIME_SEC_30D} as maximum.
-     * @see StatisticsConstants#FIELD_SEND_INTERVAL
-     */
-    long getInterval(ObjectNode json) {
-        try {
-            long interval = json.get(StatisticsConstants.FIELD_SEND_INTERVAL)
-                    .asLong();
-            return normalizeInterval(interval);
-        } catch (Exception e) {
-            // Just return the default value
-            getLogger().debug(
-                    FAILED_TO_READ + StatisticsConstants.FIELD_SEND_INTERVAL,
-                    e);
-        }
-        return StatisticsConstants.TIME_SEC_24H;
-    }
-
-    /**
-     * Gets the last time the data was collected according to the statistics
-     * file.
-     *
-     * @return Unix timestamp or -1 if not present
-     * @see StatisticsConstants#FIELD_LAST_SENT
-     */
-    long getLastSendTime(ObjectNode json) {
-        try {
-            return json.get(StatisticsConstants.FIELD_LAST_SENT).asLong();
-        } catch (Exception e) {
-            // Use default value in case of any problems
-            getLogger().debug(
-                    FAILED_TO_READ + StatisticsConstants.FIELD_LAST_SENT, e);
-        }
-        return -1; //
-    }
-
-    /**
-     * Gets the last time the data was collected according to the statistics
-     * file.
-     *
-     * @return Unix timestamp or -1 if not present
-     * @see StatisticsConstants#FIELD_LAST_STATUS
-     */
-    String getLastSendStatus(ObjectNode json) {
-        try {
-            return json.get(StatisticsConstants.FIELD_LAST_STATUS).asText();
-        } catch (Exception e) {
-            // Use default value in case of any problems
-            getLogger().debug(
-                    FAILED_TO_READ + StatisticsConstants.FIELD_LAST_STATUS, e);
-        }
-        return null; //
-    }
-
-    /**
      * Runs the given command with the store locked.
      * 
      * @param whenLocked
@@ -342,12 +85,14 @@ public class StatisticsStorage {
      * Updates the store in a safe way.
      * 
      * @param updater
-     *            the update logic
+     *            the update logic which receives a global and a project
+     *            specific container to update
      */
     void update(BiConsumer<StatisticsContainer, StatisticsContainer> updater) {
         access(() -> {
             ObjectNode fullJson = internalRead();
             ObjectNode projectJson = getProjectData(fullJson, projectId);
+
             updater.accept(new StatisticsContainer(fullJson),
                     new StatisticsContainer(projectJson));
             // TODO Is fullJson updated automatically from projectJson or not?
@@ -397,12 +142,11 @@ public class StatisticsStorage {
      *
      * @see #getUsageStatisticsFile()
      */
-    ObjectNode internalRead() {
+    private ObjectNode internalRead() {
         File file = getUsageStatisticsFile();
         getLogger().debug("Reading statistics from {}", file.getAbsolutePath());
         try {
             if (file.exists()) {
-                // Read full data and make sure we track the right project
                 return (ObjectNode) JsonHelpers.getJsonMapper().readTree(file);
             }
         } catch (JsonProcessingException e) {
@@ -411,8 +155,7 @@ public class StatisticsStorage {
             getLogger().debug("Failed to read json", e);
         }
 
-        // Empty node if nothing else is found and make sure we
-        // track the right project
+        // Empty node if nothing is found
         ObjectNode json = JsonHelpers.getJsonMapper().createObjectNode();
         json.set(StatisticsConstants.FIELD_PROJECTS,
                 JsonHelpers.getJsonMapper().createArrayNode());
@@ -447,28 +190,11 @@ public class StatisticsStorage {
         return this.usageStatisticsFile;
     }
 
-    /**
-     * Set custom usage statistics json file location.
-     *
-     * If not set <code>ProjectHelpers.resolveStatisticsStore()</code> is used.
-     *
-     * @see ProjectHelpers#resolveStatisticsStore
-     *
-     */
-    void setUsageStatisticsStore(File usageStatistics) {
-        this.usageStatisticsFile = usageStatistics;
-    }
-
-    /**
-     * Get the last server message.
-     *
-     * @return The message string returned from server in last successful
-     *         requests.
-     */
-    String getLastServerMessage(ObjectNode json) {
-        return json.has(StatisticsConstants.FIELD_SERVER_MESSAGE)
-                ? json.get(StatisticsConstants.FIELD_SERVER_MESSAGE).asText()
-                : null;
+    void clearAllProjectData() {
+        update((global, project) -> {
+            global.setValue(StatisticsConstants.FIELD_PROJECTS,
+                    JsonHelpers.getJsonMapper().createArrayNode());
+        });
     }
 
 }
